requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
asyncer==0.0.2
bidict==0.23.1
certifi==2024.6.2
chainlit==1.1.306
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.1
fastapi==0.110.3
fastapi-socketio==0.0.10
filetype==1.2.0
FLAML==2.1.2
fonttools==4.53.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
kiwisolver==1.4.5
Lazify==0.4.0
literalai==0.0.607
marshmallow==3.21.2
matplotlib==3.9.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.31.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
packaging==23.2
pillow==10.3.0
protobuf==4.25.3
pyautogen==0.2.28
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
regex==2024.5.15
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
syncer==2.0.3
termcolor==2.4.0
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.1
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
